refactor(discovery): split brief and prd out of the phase orchestrator
372 lines held phase B, phase C and the sequence they run in. The merge adds
re-entry logic to that file, so the split happens before it grows, not after.

Bodies are unchanged; the sequencer keeps only the order and the revise loop.
Tests follow the code: the write_prd cases move to their own file, and the
sequencer's tests now patch `brief.sub_tools` rather than reaching through
the orchestrator for something it no longer owns.
